+++ Sources/DeckKit/Deck.swift
@@ -14,33 +14,64 @@ import Foundation
  */
 public struct Deck<Item: DeckItem>: Identifiable, Equatable {
     
+    /**
+     Create a deck with items.
+     
+     - Parameters:
+       - id: The unique id of the deck.
+       - name: The name of the deck.
+       - items: The items to include in the deck.
+     */
     public init(
+        id: UUID = UUID(),
         name: String,
         items: [Item]) {
+        self.id = id
         self.name = name
         self.items = items
     }
     
-    public let id = UUID()
+    /**
+     The unique id of the deck.
+     */
+    public let id: UUID
+    
+    /**
+     The name of the deck.
+     */
     public let name: String
+    
+    /**
+     The items that are added to the deck.
+     */
     public var items: [Item]
 }
 
 public extension Deck {
     
-    func index(of card: Item) -> Int {
-        items.firstIndex { $0.id == card.id } ?? 0
+    /**
+     The index of a certain card item, if any.
+     */
+    func index(of card: Item) -> Int? {
+        items.firstIndex { $0.id == card.id }
     }
     
+    /**
+     Move a card to the back of the deck.
+     */
     mutating func moveToBack(_ card: Item) {
-        let topCard = items.remove(at: index(of: card))
+        guard let index = index(of: card) else { return }
+        let topCard = items.remove(at: index)
         items.append(topCard)
     }
     
+    /**
+     Move a card to the front of the deck.
+     */
     mutating func moveToFront(_ card: Item) {
-        guard items.count > 0 else { return }
+        guard let index = index(of: card) else { return }
         if items[0].id == card.id { return }
-        let topCard = items.remove(at: index(of: card))
+        let topCard = items.remove(at: index)
         items.insert(topCard, at: 0)
     }
 }