+++ Sources/DeckKit/Views/StackedDeck.swift
@@ -259,7 +259,7 @@ private extension View {
 
 struct StackedDeck_Previews: PreviewProvider {
     
-    static var item1: BasicItem { BasicItem(
+    static var item1: BasicCard.Item { BasicCard.Item(
         title: "Title 1",
         text: "Text 1",
         footnote: "Footnote 1",
@@ -267,7 +267,7 @@ struct StackedDeck_Previews: PreviewProvider {
         tintColor: .yellow)
     }
     
-    static var item2: BasicItem { BasicItem(
+    static var item2: BasicCard.Item { BasicCard.Item(
         title: "Title 2",
         text: "Text 2",
         footnote: "Footnote 2",